requirements.txt
pip==20.1.1
Flask==1.1.2
virtualenv==20.0.21
mysql-connector==2.2.9
mysqlclient==1.4.6
sqlalchemy==1.3.17
absl-py==0.9.0
pyaml==20.4.0
